Add Animation::apply, use doubles for frame numbers

+++ src/core/animation/keyframe.h
@@ -3,6 +3,6 @@
 #include "../skeleton/node.h"
 
 struct KeyFrame {
-    float frame; // The starting frame of the animation
+    double frame; // The starting frame of the animation
     RigidBodyTransform trans;
 };
